refactor(engine): rename map commands to discussion-map and discovery-map
The map verb covered two domain objects — the Discussion Map's subtopics
and the Discovery Map's ordering. Split it into sibling command groups
named after their objects: discussion-map add/set and discovery-map
sequence. Module layout follows: domain/map.cjs becomes
domain/discussion-map.cjs, sequenceMap moves out of domain/transitions.cjs
into a new domain/discovery-map.cjs, and lib.cjs exports the discussion-map
module as engine.discussionMap. No behaviour change — identical JSON
responses and manifest effects; the old map verb now falls through to the
standard usage error. Tests split to match: test-engine-discussion-map.cjs
and test-engine-discovery-map.cjs.
