+++ package.json
@@ -7,7 +7,7 @@
     "build": "node build/knowledge.build.js",
     "knowledge:repl": "node scripts/knowledge-repl.js",
     "typecheck": "tsc -p tsconfig.json",
-    "test": "node --test tests/scripts/test-render.cjs tests/scripts/test-engine-gateway.cjs tests/scripts/test-engine-epic-projections.cjs tests/scripts/test-engine-start-projections.cjs tests/scripts/test-engine-workunit-projections.cjs tests/scripts/test-engine-map.cjs tests/scripts/test-engine-tasks.cjs tests/scripts/test-engine-transactions.cjs tests/scripts/test-engine-cache.cjs tests/scripts/test-engine-boot.cjs tests/scripts/test-discovery-for-start.cjs tests/scripts/test-discovery-for-workflow-continue-epic.cjs tests/scripts/test-discovery-for-workflow-continue-feature.cjs tests/scripts/test-discovery-for-workflow-continue-bugfix.cjs tests/scripts/test-discovery-for-workflow-continue-quickfix.cjs tests/scripts/test-discovery-for-workflow-continue-cross-cutting.cjs",
+    "test": "node --test tests/scripts/test-render.cjs tests/scripts/test-engine-gateway.cjs tests/scripts/test-engine-epic-projections.cjs tests/scripts/test-engine-start-projections.cjs tests/scripts/test-engine-workunit-projections.cjs tests/scripts/test-engine-discussion-map.cjs tests/scripts/test-engine-discovery-map.cjs tests/scripts/test-engine-tasks.cjs tests/scripts/test-engine-transactions.cjs tests/scripts/test-engine-cache.cjs tests/scripts/test-engine-boot.cjs tests/scripts/test-discovery-for-start.cjs tests/scripts/test-discovery-for-workflow-continue-epic.cjs tests/scripts/test-discovery-for-workflow-continue-feature.cjs tests/scripts/test-discovery-for-workflow-continue-bugfix.cjs tests/scripts/test-discovery-for-workflow-continue-quickfix.cjs tests/scripts/test-discovery-for-workflow-continue-cross-cutting.cjs",
     "test:migrations": "for t in tests/scripts/test-migration-*.sh; do bash $t || exit 1; done"
   },
   "devDependencies": {
