+++ skills/workflow-engine/scripts/domain/discussion-map.cjs
