+++ skills/workflow-engine/scripts/domain/transitions.cjs
@@ -1,10 +1,10 @@
 'use strict';
 
 // ---------------------------------------------------------------------------
-// Domain ring: epic topic transitions — cancel, reactivate, and discovery-map
-// sequencing as single transactions. Each call is one atomic state change
-// from the caller's perspective: manifest write, knowledge-base sync (where
-// applicable), scoped git commit.
+// Domain ring: epic topic transitions — cancel and reactivate as single
+// transactions. Each call is one atomic state change from the caller's
+// perspective: manifest write, knowledge-base sync (where applicable),
+// scoped git commit.
 //
 // The manifest write is the source of truth and lands first; the knowledge
 // base is a derived index, so its failures are recorded as warnings, never
@@ -152,54 +152,4 @@ function reactivateTopic(cwd, workUnit, phase, topic) {
   return result;
 }
 
-/**
- * @typedef {object} SequenceResult
- * @property {Record<string, number>} ordered  topic → order, as applied
- * @property {string|null} committed  short commit sha, or null when nothing was staged
- * @property {string} [note]          set when committed is null
- */
-
-/**
- * Record a suggested execution order across discovery-map topics: set each
- * topic's `order`, commit scoped to the work unit. Judgment (choosing the
- * order) is the caller's; this validates and writes it. Every topic must
- * exist under `phases.discovery.items` and every order must be a positive
- * integer — checked before anything is written.
- * @param {string} cwd project root
- * @param {string} workUnit
- * @param {Record<string, number>} orders  topic → order
- * @returns {SequenceResult}
- */
-function sequenceMap(cwd, workUnit, orders) {
-  const manifest = loadWorkUnitManifest(cwd, workUnit);
-  const discovery = manifest.phases && typeof manifest.phases === 'object' ? manifest.phases.discovery : undefined;
-  const items = discovery && typeof discovery === 'object' ? discovery.items : undefined;
-  if (!items || typeof items !== 'object') {
-    throw new Error('no discovery items in the manifest (phases.discovery.items)');
-  }
-  const entries = Object.entries(orders);
-  if (entries.length === 0) {
-    throw new Error('no {topic}={order} assignments given');
-  }
-  for (const [topic, order] of entries) {
-    if (!items[topic] || typeof items[topic] !== 'object') {
-      throw new Error(`no discovery item "${topic}" in the manifest (phases.discovery.items)`);
-    }
-    if (!Number.isInteger(order) || order < 1) {
-      throw new Error(`order for "${topic}" must be a positive integer (got ${JSON.stringify(order)})`);
-    }
-  }
-  for (const [topic, order] of entries) {
-    items[topic].order = order;
-  }
-
-  saveWorkUnitManifest(cwd, workUnit, manifest);
-
-  const committed = commitScoped(cwd, `.workflows/${workUnit}`, `discovery(${workUnit}): sequence topic map`);
-  /** @type {SequenceResult} */
-  const result = { ordered: orders, committed };
-  if (committed === null) result.note = 'nothing to commit';
-  return result;
-}
-
-module.exports = { cancelTopic, reactivateTopic, sequenceMap };
+module.exports = { cancelTopic, reactivateTopic };
